Implementing select curves and updating extrusionrulesprocessor to rulesprocessor

+++ src/Apps/3dView/Model3DUtils.cpp
@@ -164,7 +164,7 @@ Eigen::Affine3f Model3DUtils::normalizePointCloud( float minx, float maxx, float
 }
 
 
-Curve2D Model3DUtils::convertToCurve2D( const std::vector< float >& v )
+Curve2D Model3DUtils::convertToCurve2D( const std::vector< double >& v )
 {
 
     Curve2D curve;
